+++ src/lib/components/FixedSizeList/FixedSizeList.tsx
@@ -1,9 +1,8 @@
 import React, { CSSProperties } from "react";
 import { useEffect, useState } from "react";
 import { RefObject } from "react";
-import useScrollDetector from "../useScrollDetector";
-import useViewportHeight from "../useViewportSize";
-import { calculateStartIndex, calculateRenderIndex } from "../utils";
+import { useScrollDetector, useViewportHeight } from "../../hooks";
+import { calculateStartIndex, calculateRenderIndex } from "../../utils";
 
 interface FixedSizeListProps {
   scrollTarget?: RefObject<any>;
