+++ src/lib/components/index.ts
@@ -1,0 +1,1 @@
+export { default as FixedSizeList } from "./FixedSizeList/FixedSizeList";
