+++ src/lib/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useScrollDetector } from "./useScrollDetector/useScrollDetector";
+export { useViewportHeight } from "./useViewportHeight/useViewportHeight";
