+++ src/lib/index.ts
@@ -1,3 +1,3 @@
-import FixedSizeList from "./components/FixedSizeList";
+import FixedSizeList from "./components/FixedSizeList/FixedSizeList";
 
 export { FixedSizeList };
