+++ src/lib/utils/calculate-index.ts
